Reorg assistant into domains/assistant
The assistant cluster relocated into domains/assistant using the proven safe
approach:
- loop.ts / data.ts / maintenance.ts (the Convex-function files) git-mv'd whole;
  root assistant.ts / assistantData.ts / assistantMaintenance.ts re-export them,
  keeping every api.assistant* / internal.assistant* path fixed
- tools.ts / history.ts (+ tests) git-mv'd from lib/ (cluster-internal, no facade)
- tables.ts + validators.ts: the four assistant tables + assistantBlockValidator
  moved out of schema.ts and composed back in
- lib/assistantLogic stays in lib/ — it's a genuinely shared helper
  (googleEventIdForOperation is imported by booking, calendar, and the Google
  adapter), so moving it into the assistant domain would invert the dependency

Behavior-preserving: types pass, `convex dev --once` deploys clean (validates the
re-exported function paths), full suite green, apps/web builds. The only api.d.ts
removals are the two non-function helper modules (lib/assistantTools,
lib/assistantHistory) that relocated — no callable path changed.
