+++ packages/backend/convex/_generated/api.d.ts
@@ -18,6 +18,13 @@ import type * as calendar from "../calendar.js";
 import type * as calendarSync from "../calendarSync.js";
 import type * as contacts from "../contacts.js";
 import type * as crons from "../crons.js";
+import type * as domains_assistant_data from "../domains/assistant/data.js";
+import type * as domains_assistant_history from "../domains/assistant/history.js";
+import type * as domains_assistant_loop from "../domains/assistant/loop.js";
+import type * as domains_assistant_maintenance from "../domains/assistant/maintenance.js";
+import type * as domains_assistant_tables from "../domains/assistant/tables.js";
+import type * as domains_assistant_tools from "../domains/assistant/tools.js";
+import type * as domains_assistant_validators from "../domains/assistant/validators.js";
 import type * as domains_booking_model from "../domains/booking/model.js";
 import type * as domains_booking_mutations from "../domains/booking/mutations.js";
 import type * as domains_booking_queries from "../domains/booking/queries.js";
@@ -47,9 +54,7 @@ import type * as integrations_calendar_service from "../integrations/calendar/se
 import type * as integrations_calendar_types from "../integrations/calendar/types.js";
 import type * as integrations_google_adapter from "../integrations/google/adapter.js";
 import type * as integrations_google_mappers from "../integrations/google/mappers.js";
-import type * as lib_assistantHistory from "../lib/assistantHistory.js";
 import type * as lib_assistantLogic from "../lib/assistantLogic.js";
-import type * as lib_assistantTools from "../lib/assistantTools.js";
 import type * as lib_calendars from "../lib/calendars.js";
 import type * as lib_eventReads from "../lib/eventReads.js";
 import type * as lib_google from "../lib/google.js";
@@ -79,6 +84,13 @@ declare const fullApi: ApiFromModules<{
   calendarSync: typeof calendarSync;
   contacts: typeof contacts;
   crons: typeof crons;
+  "domains/assistant/data": typeof domains_assistant_data;
+  "domains/assistant/history": typeof domains_assistant_history;
+  "domains/assistant/loop": typeof domains_assistant_loop;
+  "domains/assistant/maintenance": typeof domains_assistant_maintenance;
+  "domains/assistant/tables": typeof domains_assistant_tables;
+  "domains/assistant/tools": typeof domains_assistant_tools;
+  "domains/assistant/validators": typeof domains_assistant_validators;
   "domains/booking/model": typeof domains_booking_model;
   "domains/booking/mutations": typeof domains_booking_mutations;
   "domains/booking/queries": typeof domains_booking_queries;
@@ -108,9 +120,7 @@ declare const fullApi: ApiFromModules<{
   "integrations/calendar/types": typeof integrations_calendar_types;
   "integrations/google/adapter": typeof integrations_google_adapter;
   "integrations/google/mappers": typeof integrations_google_mappers;
-  "lib/assistantHistory": typeof lib_assistantHistory;
   "lib/assistantLogic": typeof lib_assistantLogic;
-  "lib/assistantTools": typeof lib_assistantTools;
   "lib/calendars": typeof lib_calendars;
   "lib/eventReads": typeof lib_eventReads;
   "lib/google": typeof lib_google;
