+++ packages/backend/convex/schema.ts
@@ -1,6 +1,7 @@
 import { defineSchema, defineTable } from "convex/server";
 import { v } from "convex/values";
 
+import { assistantTables } from "./domains/assistant/tables";
 import { bookingTables } from "./domains/booking/tables";
 import { calendarTables } from "./domains/calendar/tables";
 import { marketingTables } from "./domains/marketing/tables";
@@ -11,41 +12,7 @@ import { peopleTables } from "./domains/people/tables";
 // domains/calendar/validators.ts so the schema and the calendar write path can
 // share them without a circular import back through this file.
 
-/** One piece of an assistant turn, in the order it happened.
- *
- * A turn is a list of these rather than a string because a single reply can
- * interleave prose with tool activity, and the panel renders each kind
- * differently. The action appends to the list as the model streams, so the same
- * shape has to survive a half-finished turn.
- *
- * `tool_call` and `tool_result` are also what the next request's history is
- * rebuilt from, so they hold exactly what the model needs to see: the raw
- * argument JSON as it arrived, never a re-serialized version of it.
- *
- * `proposal` is the odd one out — it carries nothing for the model, only the id
- * of the `assistantActions` row the panel renders a confirm card for. */
-export const assistantBlockValidator = v.union(
-  v.object({ type: v.literal("text"), text: v.string() }),
-  v.object({
-    type: v.literal("tool_call"),
-    toolCallId: v.string(),
-    name: v.string(),
-    // The model's own JSON string. Kept verbatim: re-encoding it would change
-    // the bytes the model sees when this turn is replayed as history.
-    arguments: v.string(),
-  }),
-  v.object({
-    type: v.literal("tool_result"),
-    toolCallId: v.string(),
-    content: v.string(),
-    isError: v.optional(v.boolean()),
-  }),
-  v.object({
-    type: v.literal("proposal"),
-    toolCallId: v.string(),
-    actionId: v.id("assistantActions"),
-  }),
-);
+// assistantBlockValidator lives in domains/assistant/validators.ts.
 
 export default defineSchema({
   // One row per user tracking incremental-sync state for Google data.
@@ -117,100 +84,9 @@ export default defineSchema({
   // These tables stay empty when no DEEPSEEK_API_KEY is configured; the
   // assistant is optional and its absence must not change anything else.
 
-  // One conversation. `lastMessageAt` rather than `_creationTime` orders the
-  // list, so a revived old thread sorts to the top where the user left it.
-  assistantThreads: defineTable({
-    userId: v.string(),
-    title: v.string(),
-    createdAt: v.number(),
-    lastMessageAt: v.number(),
-    // Transactional per-thread turn claim. Cleared by finish/fail; stale claims
-    // are recovered against the server timestamp when the next turn starts.
-    activeMessageId: v.optional(v.id("assistantMessages")),
-  }).index("by_user_and_lastMessage", ["userId", "lastMessageAt"]),
-
-  // One bounded operational row per assistant user: fixed-window request quota
-  // and a global one-turn lease. Keeping this separate avoids churning threads.
-  assistantUserState: defineTable({
-    userId: v.string(),
-    windowStartMs: v.number(),
-    requestCount: v.number(),
-    // Rolling monthly quota, independent of the 5-minute burst window above.
-    monthWindowStartMs: v.optional(v.number()),
-    monthCount: v.optional(v.number()),
-    activeMessageId: v.optional(v.id("assistantMessages")),
-    activeThreadId: v.optional(v.id("assistantThreads")),
-    leaseExpiresAt: v.optional(v.number()),
-  }).index("by_user", ["userId"]),
-
-  // One row per turn. `blocks` is appended to in place while the model streams,
-  // so the client's reactive subscription renders the reply as it arrives —
-  // Convex actions can't stream to the browser, but a patched row does.
-  assistantMessages: defineTable({
-    threadId: v.id("assistantThreads"),
-    // Denormalized from the thread so every read can be scoped without a join.
-    userId: v.string(),
-    role: v.union(v.literal("user"), v.literal("assistant")),
-    blocks: v.array(assistantBlockValidator),
-    // A turn that is still streaming is renderable but not yet replayable as
-    // history; `error` holds why a turn stopped, for the panel to show.
-    status: v.union(
-      v.literal("streaming"),
-      v.literal("complete"),
-      v.literal("error"),
-    ),
-    error: v.optional(v.string()),
-    // Suggested next prompts, generated best-effort once a turn settles and only
-    // when they'd genuinely help — usually absent. The panel renders them as
-    // clickable chips under the latest reply.
-    suggestions: v.optional(v.array(v.string())),
-    createdAt: v.number(),
-  })
-    .index("by_thread", ["threadId", "createdAt"])
-    // userId is denormalized here for user-scoped reads; the index also lets
-    // account-deletion cleanup remove a user's messages without a thread join.
-    .index("by_user", ["userId"]),
-
-  // A write the assistant wants to make, held until the user confirms it.
-  //
-  // The assistant's tools never reach Google. They record one of these instead,
-  // and only `assistant.confirmAction` — driven by a click — applies it. That
-  // makes this table both the confirmation queue and the permanent audit trail
-  // of everything the assistant ever proposed, applied, or was refused.
-  assistantActions: defineTable({
-    threadId: v.id("assistantThreads"),
-    userId: v.string(),
-    // Ties the proposal back to the tool call that produced it.
-    toolCallId: v.string(),
-    tool: v.string(),
-    // The model's argument JSON. Re-validated against the tool's own schema at
-    // apply time — never trusted just because it was stored.
-    input: v.string(),
-    // A one-line description of the change, written when the proposal is made.
-    // The confirm card renders this, so what the user approves is stated in
-    // words rather than inferred from raw arguments.
-    preview: v.string(),
-    // Stable across retries and used as Google's client-selected event ID.
-    operationId: v.optional(v.string()),
-    attemptCount: v.optional(v.number()),
-    applyLeaseExpiresAt: v.optional(v.number()),
-    // `applying` is the claim a confirm click takes before it calls Google, so
-    // a double-click can't send the same invitation twice: the second click
-    // finds the row already out of `pending` and does nothing.
-    status: v.union(
-      v.literal("pending"),
-      v.literal("applying"),
-      v.literal("applied"),
-      v.literal("rejected"),
-      v.literal("failed"),
-    ),
-    // What happened on apply: a human-readable confirmation, or the error.
-    resultSummary: v.optional(v.string()),
-    createdAt: v.number(),
-    decidedAt: v.optional(v.number()),
-  })
-    .index("by_thread", ["threadId", "createdAt"])
-    .index("by_user_and_status", ["userId", "status"]),
+  // Assistant domain tables — threads / state / messages / actions (see
+  // domains/assistant/tables.ts).
+  ...assistantTables,
 
   // Marketing domain tables — the public waitlist (see domains/marketing/tables.ts).
   ...marketingTables,
